+++ lib/src/push.dart
@@ -1,14 +1,19 @@
 import 'dart:async';
 
-import 'package:equatable/equatable.dart';
 import 'package:logging/logging.dart';
-import 'package:phoenix_socket/phoenix_socket.dart';
-import 'package:quiver/collection.dart';
+
+import 'channel.dart';
+import 'events.dart';
+import 'exceptions.dart';
+import 'message.dart';
+import 'version.dart';
+
+typedef ReceiverCallback = void Function(PushResponse response);
 
 /// Encapsulates the response to a [Push].
-class PushResponse extends Equatable {
+class PushResponse {
   /// Builds a PushResponse from a status and response.
-  PushResponse({
+  const PushResponse({
     this.status,
     this.response,
   });
@@ -27,10 +32,10 @@ class PushResponse extends Equatable {
   /// }
   /// ```
   factory PushResponse.fromMessage(Message message) {
-    final data = message.payload!;
+    final data = message.payload;
     return PushResponse(
-      status: data['status'] as String?,
-      response: data['response'],
+      status: data?['status'] as String?,
+      response: data?['response'],
     );
   }
 
@@ -52,10 +57,16 @@ class PushResponse extends Equatable {
   bool get isTimeout => status == 'timeout';
 
   @override
-  List<Object?> get props => [status, response];
+  bool operator ==(Object other) =>
+      other is PushResponse &&
+      other.status == status &&
+      other.response == response;
 
   @override
-  bool get stringify => true;
+  int get hashCode => Object.hash(status, response);
+
+  @override
+  String toString() => 'PushResponse(status: $status, response: $response)';
 }
 
 /// Type of function that should return a push payload
@@ -79,8 +90,7 @@ class Push {
         _responseCompleter = Completer<PushResponse>();
 
   final Logger _logger;
-  final ListMultimap<String, void Function(PushResponse)> _receivers =
-      ListMultimap();
+  final Map<String, List<ReceiverCallback>> _receivers = {};
 
   /// The event name associated with the pushed message
   final PhoenixChannelEvent? event;
@@ -160,7 +170,7 @@ class Push {
       // ignore: avoid_catches_without_on_clauses
     } catch (err, stacktrace) {
       _logger.warning(
-        'Catched error for push $ref',
+        'Caught error for push $ref',
         err,
         stacktrace,
       );
@@ -174,17 +184,16 @@ class Push {
   /// after a reconnection.
   Future<void> resend(Duration? newTimeout) async {
     timeout = newTimeout ?? timeout;
-    reset();
+    if (_sent) {
+      reset();
+    }
     await send();
   }
 
   /// Associate a callback to be called if and when a reply with the given
   /// status is received.
-  void onReply(
-    String status,
-    void Function(PushResponse) callback,
-  ) {
-    _receivers[status].add(callback);
+  void onReply(String status, ReceiverCallback callback) {
+    (_receivers[status] ??= []).add(callback);
   }
 
   /// Schedule a timeout to be triggered if no reply occurs
@@ -200,7 +209,7 @@ class Push {
 
     _timeoutTimer ??= Timer(timeout!, () {
       _timeoutTimer = null;
-      _logger.warning('Push $ref timed out');
+      _logger.warning(() => 'Push $ref timed out');
       _channel.trigger(Message.timeoutFor(ref, _version));
     });
   }
@@ -246,13 +255,14 @@ class Push {
     }
 
     _logger.finer(() {
-      if (_receivers[response.status].isNotEmpty) {
-        return 'Triggering ${_receivers[response.status].length} callbacks';
+      if (_receivers[response.status] case final receiver?
+          when receiver.isNotEmpty) {
+        return 'Triggering ${receiver.length} callbacks';
       }
       return 'Not triggering any callbacks';
     });
 
-    final receivers = _receivers[response.status].toList();
+    final receivers = _receivers[response.status]?.toList() ?? const [];
     clearReceivers();
     for (final cb in receivers) {
       cb(response);
@@ -260,14 +270,15 @@ class Push {
   }
 
   /// Dispose the set of waiters associated with this push.
-  void clearReceivers() {
-    _receivers.clear();
-  }
+  void clearReceivers() => _receivers.clear();
 
   // Remove existing waiters and reset completer
   void cleanUp() {
-    clearReceivers();
-    _responseCompleter = Completer();
+    if (_sent) {
+      _logger.fine('Cleaning up completer');
+      clearReceivers();
+      _responseCompleter = Completer();
+    }
   }
 
   void _receiveResponse(dynamic response) {
@@ -276,7 +287,9 @@ class Push {
       if (response.event == replyEvent) {
         trigger(PushResponse.fromMessage(response));
       }
-    } else if (response is PhoenixException) {
+    } else if (event != PhoenixChannelEvent.join) {
+      _logger.finest(
+          () => "Completing with error: ${_responseCompleter.hashCode}");
       if (!_responseCompleter.isCompleted) {
         _responseCompleter.completeError(response);
         clearReceivers();